+++ R/simYes.r
@@ -29,7 +29,7 @@ simYes.opiQueryDevice   <- function() { return (list(type="SimNo", isSim=TRUE)) 
 #
 # Return NULL if succesful, string error message otherwise  
 ################################################################################
-simYes.opiInitialize <- function(display=NULL) {
+simYes.opiInitialize <- function(display = NA) {
     if(simDisplay.setupDisplay(display))
         warning("opiInitialize (SimNo): display parameter may not contain 4 numbers.")
 
@@ -42,14 +42,25 @@ simYes.opiInitialize <- function(display=NULL) {
 #   NULL - succsess
 #   -1   - opiInitialize not called
 ################################################################################
+#' @rdname opiSetBackground
+#' @details
+#' \subsection{SimYes}{
+#'   DETAILS
+#' }
 simYes.opiSetBackground <- function(col, gridCol) { 
     return (simDisplay.setBackground(col, gridCol))
 }
 
-################################################################################
-#
-################################################################################
-simYes.opiPresent <- function(stim) {
+#' @rdname opiPresent
+#' @details
+#' \subsection{SimYes}{
+#'   \code{opiPresent(stim, nextStim=NULL)}
+#'   
+#'   If the chosen OPI implementation is \code{SimYes}, then the response to a
+#'   stimuli is always yes, hence \code{\link{opiPresent}} always returns
+#'   \code{err=NULL}, \code{seen=TRUE}, and \code{time=0}.
+#' }
+simYes.opiPresent <- function(stim, nextStim=NULL) {
     simDisplay.present(stim$x, stim$y, stim$color, stim$duration, stim$responseWindow)
 
     return ( list(